feat(enquete): refactor enquete validation activite
