+++ packages/api/src/routes/hasura-actions/enquete/common/enquete-status/activite/enqueteMandataireIndividuelActiviteStatus.js
@@ -1,5 +1,8 @@
 const { buildMesureGroupsAttributes } = require("./activiteSchemaUtil");
-const { getValidationStatus, getGlobalStatus } = require("../../common");
+const {
+  getValidationStatus,
+  getGlobalStatus
+} = require("../enqueteSchemaUtil");
 const yup = require("yup");
 
 async function enqueteMandataireIndividuelActiviteStatus(enqueteReponse) {
