+++ packages/api/src/routes/hasura-actions/enquete/common/enquete-status/index.js
@@ -1,0 +1,8 @@
+const { getGlobalStatus, getValidationStatus } = require("./enqueteSchemaUtil");
+const enqueteMandataireIndividuelActiviteStatus = require("./enqueteMandataireIndividuelActiviteStatus");
+
+module.exports = {
+  getGlobalStatus,
+  getValidationStatus,
+  enqueteMandataireIndividuelActiviteStatus
+};
