--- packages/api/src/routes/hasura-actions/enquete/common/index.js
@@ -1,6 +1,0 @@
-const { getGlobalStatus, getValidationStatus } = require("./enqueteSchemaUtil");
-
-module.exports = {
-  getGlobalStatus,
-  getValidationStatus
-};
