--- packages/api/src/routes/hasura-actions/enquete/mandataire-individuel-status/activite/index.js
@@ -1,1 +1,0 @@
-module.exports = require("./enqueteMandataireIndividuelActiviteStatus");
